requirements.txt
streamlit
openai
python-docx
PyPDF2
requests
pandas
openpyxl